+++ optaplanner-examples/src/main/java/org/optaplanner/examples/vehiclerouting/domain/location/AirLocation.java
@@ -33,10 +33,10 @@ public class AirLocation extends Location {
     }
 
     @Override
-    public int getDistanceTo(Location location) {
+    public long getDistanceTo(Location location) {
         double distance = getAirDistanceDoubleTo(location);
         // Multiplied by 1000 to avoid floating point arithmetic rounding errors
-        return (int) (distance * 1000.0 + 0.5);
+        return (long) (distance * 1000.0 + 0.5);
     }
 
 }
